chore: Drop codemirror-shim as no longer needed
