--- packages/codemirror-shim/package.json
@@ -1,51 +1,0 @@
-{
-    "name": "@hpcc-js/codemirror-shim",
-    "version": "3.0.0",
-    "description": "hpcc-js - codemirror shim",
-    "type": "module",
-    "exports": {
-        ".": {
-            "types": "./types/index.d.ts",
-            "default": "./dist/index.js"
-        },
-        "./dist/*": "./dist/*"
-    },
-    "module": "./dist/index.js",
-    "browser": "./dist/index.js",
-    "types": "./types/index.d.ts",
-    "files": [
-        "dist/*",
-        "src/*",
-        "types/*"
-    ],
-    "scripts": {
-        "clean": "rimraf --glob lib* types dist *.tsbuildinfo .turbo",
-        "bundle": "node esbuild.js",
-        "bundle-watch": "npm run bundle -- --development --watch",
-        "gen-types": "tsc --project tsconfig.json",
-        "gen-types-watch": "npm run gen-types -- --watch",
-        "build": "run-p gen-types bundle",
-        "stamp": "node ../../node_modules/@hpcc-js/bundle/src/stamp.js",
-        "lint": "eslint ./src",
-        "lint-fix": "eslint --fix src/**/*.ts",
-        "docs": "typedoc --options tdoptions.json .",
-        "update": "npx --yes npm-check-updates -u -t minor",
-        "update-major": "npx --yes npm-check-updates -u"
-    },
-    "dependencies": {},
-    "devDependencies": {
-        "@hpcc-js/esbuild-plugins": "^1.2.0",
-        "codemirror": "5.65.18"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/hpcc-systems/Visualization.git"
-    },
-    "author": "Gordon Smith <gordonjsmith@gmail.com>",
-    "contributors": [],
-    "license": "Apache-2.0",
-    "bugs": {
-        "url": "https://github.com/hpcc-systems/Visualization/issues"
-    },
-    "homepage": "https://github.com/hpcc-systems/Visualization"
-}
