--- packages/codemirror-shim/src/__package__.ts
@@ -1,3 +1,0 @@
-export const PKG_NAME = "@hpcc-js/codemirror-shim";
-export const PKG_VERSION = "2.37.0";
-export const BUILD_VERSION = "2.107.0";
