--- packages/codemirror-shim/tsconfig.json
@@ -1,24 +1,0 @@
-{
-    "compilerOptions": {
-        "rootDir": "src",
-        "module": "NodeNext",
-        "target": "ESNext",
-        "resolveJsonModule": true,
-        "emitDeclarationOnly": true,
-        "declaration": true,
-        "declarationDir": "types",
-        "strict": true,
-        "noImplicitAny": false,
-        "strictNullChecks": false,
-        "skipLibCheck": true,
-        "allowImportingTsExtensions": true,
-        "lib": [
-            "DOM",
-            "ESNext",
-            "ES2020"
-        ]
-    },
-    "include": [
-        "./src/index.ts"
-    ]
-}
